+++ lib/rpush/configuration.rb
@@ -1,24 +1,40 @@
 require 'pathname'
 
 module Rpush
-  def self.config
-    @config ||= Rpush::Configuration.new
-  end
+  class << self
+    attr_writer :config
 
-  def self.configure
-    if block_given?
+    def config
+      @config ||= Rpush::Configuration.new
+    end
+
+    def configure
+      return unless block_given?
       yield config
       config.initialize_client
     end
   end
 
-  CURRENT_ATTRS = [:push_poll, :feedback_poll, :embedded, :pid_file, :batch_size,
-                   :push, :client, :logger, :log_file, :foreground, :log_level, :plugin]
-  DEPRECATED_ATTRS = [:log_dir]
+  CURRENT_ATTRS = [:push_poll, :embedded, :pid_file, :batch_size, :push, :client, :logger, :log_file, :foreground, :log_level, :plugin, :apns]
+  DEPRECATED_ATTRS = [:log_dir, :feedback_poll]
   CONFIG_ATTRS = CURRENT_ATTRS + DEPRECATED_ATTRS
 
   class ConfigurationError < StandardError; end
   class ConfigurationWithoutDefaults < Struct.new(*CONFIG_ATTRS); end # rubocop:disable Style/StructInheritance
+
+  class ApnsFeedbackReceiverConfiguration < Struct.new(:frequency) # rubocop:disable Style/StructInheritance
+    def initialize
+      super
+      self.frequency = 60
+    end
+  end
+
+  class ApnsConfiguration < Struct.new(:feedback_receiver) # rubocop:disable Style/StructInheritance
+    def initialize
+      super
+      self.feedback_receiver = ApnsFeedbackReceiverConfiguration.new
+    end
+  end
 
   class Configuration < Struct.new(*CONFIG_ATTRS) # rubocop:disable Style/StructInheritance
     include Deprecatable
@@ -29,7 +45,21 @@ module Rpush
 
     def initialize
       super
-      set_defaults
+
+      self.push_poll = 2
+      self.batch_size = 100
+      self.logger = nil
+      self.log_file = 'log/rpush.log'
+      self.pid_file = 'tmp/rpush.pid'
+      self.log_level = (defined?(Rails) && Rails.logger) ? Rails.logger.level : ::Logger::Severity::DEBUG
+      self.plugin = OpenStruct.new
+      self.foreground = false
+
+      self.apns = ApnsConfiguration.new
+
+      # Internal options.
+      self.embedded = false
+      self.push = false
     end
 
     def update(other)
@@ -68,31 +98,21 @@ module Rpush
       Modis.redis_options = options if client == :redis
     end
 
-    def set_defaults
-      self.push_poll = 2
-      self.feedback_poll = 60
-      self.batch_size = 100
-      self.logger = nil
-      self.log_file = 'log/rpush.log'
-      self.pid_file = 'tmp/rpush.pid'
-      self.log_level = (defined?(Rails) && Rails.logger) ? Rails.logger.level : ::Logger::Severity::DEBUG
-      self.plugin = OpenStruct.new
-      self.foreground = false
-
-      # Internal options.
-      self.embedded = false
-      self.push = false
+    def feedback_poll=(frequency)
+      apns.feedback_receiver.frequency = frequency
     end
+    deprecated(:feedback_poll=, '2.5.0', 'Please use apns.feedback_receiver.frequency= instead.')
 
     def initialize_client
       return if @client_initialized
       raise ConfigurationError, 'Rpush.config.client is not set.' unless client
       require "rpush/client/#{client}"
+
       client_module = Rpush::Client.const_get(client.to_s.camelize)
-      Rpush.send(:include, client_module)
+      Rpush.send(:include, client_module) unless Rpush.ancestors.include?(client_module)
 
       [:Apns, :Gcm, :Wpns, :Adm].each do |service|
-        Rpush.const_set(service, client_module.const_get(service))
+        Rpush.const_set(service, client_module.const_get(service)) unless Rpush.const_defined?(service)
       end
 
       @client_initialized = true