+++ spec/spec_helper.rb
@@ -54,7 +54,7 @@ def after_example_cleanup
   Rpush.logger = nil
   Rpush::Daemon.store = nil
   Rpush::Deprecation.muted do
-    Rpush.config.set_defaults if Rpush.config.is_a?(Rpush::Configuration)
+    Rpush.config = nil
     Rpush.config.client = RPUSH_CLIENT
   end
   Rpush.plugins.values.each(&:unload)